Implement Hash Functions Families
1. Implement families tester
2. Implement trivial hash function family
3. Implement hash function families from polynom
4. Implment gradually increase hash functions familiy
5. Implement tabulation hash functions family
6. Implement twisted tabulation hash functions family
7. Implement Nisan's PRG

+++ src/VectorUtils.cpp
@@ -160,7 +160,9 @@ std::string VectorUtils::to_string(VectorXi vector) {
     return result;
 }
 
-std::vector<int> VectorUtils::eigen_vector_to_std_vector(VectorXi eigen_vector) {
-    std::vector<int> std_vector(eigen_vector.data(), eigen_vector.data() + eigen_vector.size());
+std::vector<long> VectorUtils::eigen_vector_to_std_vector(VectorXi eigen_vector) {
+    std::vector<long> std_vector(eigen_vector.data(), eigen_vector.data() + eigen_vector.size());
     return std_vector;
 }
+
+
